refactored reducers to individual files
