+++ 16-Redux/src/components/Header.js
@@ -1,4 +1,4 @@
-import { authActions } from "../store";
+import { authActions } from "../store/auth";
 import classes from "./Header.module.css";
 import { useDispatch, useSelector } from "react-redux"; // import from react-dedux.
